requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
pytz==2024.1
requests==2.32.3
rsa==4.9
six==1.16.0
soupsieve==2.5
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
python-dotenv==1.0.1
supabase==2.6.0
scrapy==2.11.2